Several changes made to better accommodate versioning automation.
The application metadata is stored (mostly) in one place- that being screendimmer.desktop. It contains repetitive information that I previously hardcoded in different modules, such as the name or the version number. The setup script, setup.py, depends on some of this file for the app setup name and version.

utils.py has been expanded upon to retrieve the correct .desktop file path. Also contains a very useful conditional, def running_from_pycharm(), to determine the system environment running the program. It's for personal convenience to work on program features as efficiently as possible. Omission of this conditional will lead to pathing errors and configuration key errors. This is due to unique system pathing depending on how the program is running. Please consult the docstrings for my motivations.