Versioning bug fixes and quality of life improvement implemented.
A typo was removed from utils.py, where the Linux path to the .desktop file was incorrectly specified.

Retrieval of metadata was slightly adjusted. The program accounts for different pathing that changes between the build process (python setup.py build) and running as an installation. The differences are minute but make a huge a difference in program flow. See the new module configappdata.py. Refer to the docstring for my conclusion.

tray.py and setup.py retrieve app metadata from the same module, but in a slightly different way. The method to retrieve the configuration file allows for an argument that asks the user how they want to retrieve the file. Refer to both modules for more info.

In brightness.ini, changed the program brightness level to 1.0 to match the session's default brightness level.