Initial commit for screendimmer rewrite.
The goal of this rewrite is to improve general architecture and
provide ease of developer maintenance.

The main codebase is removed and replaced with a small prototype.
This prototype contains functionality only accessible via direct
code changes.

It's a proof of concept of a better approach to general architecture
and developer intuition when working with the codebase.

Run: python __main__.py in console to retrieve your setup information.

At the moment there is currently no GUI or CLI functionality. The
only way to test functions in this prototype is via direct code changes.