Created callbacks in place of commands. Added new config.py files.
This should have been two separate commits.

Due to limitations and differences with unique tk GUI elements,
I opted to use trace_add as callbacks to detect changes as they
are made to the pyvars. This was the path of least resistence,
further supported by several answers found on stack overflow.

Added a new config file for colors. Contains a variable called
'dark_mode' that will eventually be replaced with a value parsed
from a config parser. The other config file being parsed is
brightness.ini and now includes such a variable so that users
will be to optionally select with colors they want.