Finalized easy-to-use coloring infrastructure / theme (light and dark)
The supporting structure for developers to use a theme is as simple as
setting a boolean value. Not yet editable by the GUI, but still easy
enough to switch to config.py and manually switched the dark_mode value.

There are still some elements that need to be taken care of manually for
both dark mode and light mode (such as border colors, spinbox button colors,
etc). Will also look into finding a more optimal dark mode color.