Extensive refactor of colors.py and introduction of preferences.py
The most critical commit so far. Refactored colors.py by introducing
subclasses instead of nested classes. Now colors.py is exclusively
responsible for defining color constants. You can simply return
one of the subclasses to act as a theme object. (It would make the
most sense if that was done in a different module)

The new module preferences.py r/w/x's the config.ini file. Generally,
it describes the state of the configuration in config.ini. Some
info is pulled from colors.py to make this possible (such as
confirming that the current theme is dark mode, for example)