Finalized behavior with configfile, major refactors, fixed pref's bugs
Decided there was no (clean) way to avoid calling for the configuration
file multiple times. This was done for readability purposes and ease of
maintenance.

Various major refactors throughout files. For example, moved the function
that retrieves all themes into the utils file - it is not a colors.py
constant, nor a preferences.py preference. At most it would be a gui.py
utility, but I'm trying to avoid utility functions in that module, and
furthermore, doing so would result in a circular import due to where
I need that information.

Fixed preference bugs that resulted in them not being saved correctly
and not being displayed correctly. Simple forgot to rename certain
variables after re-using certain codes.

Also, the __main__.py will now set brightnesses on monitor brightness
values saved to the config in config.ini (or the default of 100), as well
as executing the user's preferences on program exit (save on exit, restore
brightnesses on exit).